requirements.txt
matplotlib==3.0.0
numpy==1.15.2
Keras==2.2.4
tensorflow==1.10.0
opencv_python==4.0.0.21
scikit_image==0.15.0
imageio==2.9.0
Pillow==8.0.1
skimage==0.0
